Move comparison tests to comparion.rs
